Move collection monkey patching to test suite. Add compat import for moved EnvironmentVarGuard
